+++ src/System/Integrate/Providers/IntegrateServiceProvider.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace System\Integrate\Providers;
+
+use System\Integrate\ServiceProvider;
+
+class IntegrateServiceProvider extends ServiceProvider
+{
+    public function register()
+    {
+        // application
+    }
+}