+++ src/System/Integrate/ServiceProvider.php
@@ -4,12 +4,28 @@ declare(strict_types=1);
 
 namespace System\Integrate;
 
+use System\Container\Container;
+
 abstract class ServiceProvider
 {
-    /** @var array Class regiter */
+    /** @var Application */
+    protected $app;
+
+    /** @var array Class register */
     protected $register = [
         // register
     ];
+
+    /**
+     * Create a new service provider instance.
+     *
+     * @param Application|System\Container\Container $app
+     * @return void
+     */
+    public function __construct(Container $app)
+    {
+        $this->app = $app;
+    }
 
     /**
      * Boot provider.
@@ -18,4 +34,10 @@ abstract class ServiceProvider
     {
         // boot
     }
+
+    public function register()
+    {
+        // register application container
+    }
+
 }